requirements.txt
kubernetes==12.0.1
python-dotenv==0.19.0
Flask==3.1.0
matplotlib==3.9.2